saved a new user to database via postman but still haven't figured out password hashing
